+++ script/clang-format
@@ -2,15 +2,18 @@
 
 from __future__ import print_function
 
+import argparse
 import multiprocessing
 import os
 import re
 import subprocess
 import sys
-
-import argparse
-import click
 import threading
+
+import click
+
+sys.path.append(os.path.dirname(__file__))
+from helpers import basepath, get_output, walk_files, filter_changed
 
 is_py2 = sys.version[0] == '2'
 
@@ -48,44 +51,6 @@ def progress_bar_show(value):
     if value is None:
         return ''
     return value
-
-
-def walk_files(path):
-    for root, _, files in os.walk(path):
-        for name in files:
-            yield os.path.join(root, name)
-
-
-def get_output(*args):
-    proc = subprocess.Popen(args, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    output, err = proc.communicate()
-    return output.decode('utf-8')
-
-
-def splitlines_no_ends(string):
-    return [s.strip() for s in string.splitlines()]
-
-
-def filter_changed(files):
-    for remote in ('upstream', 'origin'):
-        command = ['git', 'merge-base', '{}/dev'.format(remote), 'HEAD']
-        try:
-            merge_base = splitlines_no_ends(get_output(*command))[0]
-            break
-        except:
-            pass
-    else:
-        return files
-    command = ['git', 'diff', merge_base, '--name-only']
-    changed = splitlines_no_ends(get_output(*command))
-    changed = {os.path.relpath(f, os.getcwd()) for f in changed}
-    print("Changed Files:")
-    files = [p for p in files if p in changed]
-    for p in files:
-        print("  {}".format(p))
-    if not files:
-        print("  No changed files")
-    return files
 
 
 def main():
